requirements.txt
streamlit
python-dotenv
langchain
langchain_community
google-generativeai
langchain_google_genai